Added panel design
Added a panel design in PDF format. Note: this is a single PDF that contains the panel design for both the 2HP and 4HP versions, and is missing the drill hole for the output jack. It also does not have the 2-sided aspect (where if a user takes off the panel and flips the module, the CV output is at the top vs the bottom). Joe made that part directly in Eagle.